# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoMaster API",
        "version": "0.1"
    },
    "paths": {
        "/api/task/{id}/attachment": {
            "post": {
                "tags": [
                    "AttachFileToTask"
                ],
                "summary": "Attach file to task",
                "description": "Attach file to task",
                "operationId": "AttachFile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the task",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Attachment",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Attachment"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User's name",
                                        "type": "text",
                                        "example": "John Doe",
                                        "nullable": "false"
                                    },
                                    "email": {
                                        "description": "User's email",
                                        "type": "text",
                                        "example": "john.doe@todo.com",
                                        "nullable": "false"
                                    },
                                    "password_confirm": {
                                        "description": "Confirm password",
                                        "type": "text",
                                        "example": "john.doe@todo.com",
                                        "nullable": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "An error message",
                                            "type": "boolean",
                                            "example": "true",
                                            "nullable": "true"
                                        },
                                        "message": {
                                            "description": "Message result when registerting a user",
                                            "type": "string",
                                            "example": "User Created Successfully",
                                            "nullable": "true"
                                        },
                                        "token": {
                                            "description": "Access token for the newly registered user",
                                            "type": "string",
                                            "example": "4|laravel_sanctum_xabYxNyN6P77sTNv7cIW43eWZSBFxOKfYbnTHg9R0e5d5563",
                                            "nullable": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Login failed"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Login",
                "description": "User login",
                "operationId": "Login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User's email",
                                        "type": "text",
                                        "example": "john.doe@todo.com",
                                        "nullable": "false"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "text",
                                        "example": "@veryHardPW!^",
                                        "nullable": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Access token for the newly registered user",
                                            "type": "string",
                                            "example": "4|laravel_sanctum_xabYxNyN6P77sTNv7cIW43eWZSBFxOKfYbnTHg9R0e5d5563",
                                            "nullable": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Login failed"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Attachment": {
                "properties": {
                    "id": {
                        "description": "Attachment identifier",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    },
                    "filename": {
                        "description": "Attachments filename",
                        "type": "string",
                        "example": "test-attachment.docx",
                        "nullable": "true"
                    },
                    "type": {
                        "description": "file type extension of the attachment",
                        "type": "string",
                        "example": "john.doe@todo.com",
                        "nullable": "false"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}